+++ lib/gumdrop.rb
@@ -7,7 +7,14 @@ DEFAULT_OPTIONS= {
   :relative_paths => true,
   :auto_run => false,
   :force_reload => false,
-  :root => "."
+  :root => ".",
+  :log_level => :info
+}
+
+LOG_LEVELS = {
+  :info => 0,
+  :warning => 1,
+  :error => 2
 }
 
 module Gumdrop
@@ -15,9 +22,9 @@ module Gumdrop
   autoload :Context, "gumdrop/context"
   autoload :Content, "gumdrop/content"
   autoload :DeferredLoader, "gumdrop/deferred_loader"
+  autoload :DSL, "gumdrop/dsl"
   autoload :Generator, "gumdrop/generator"
   autoload :GeneratedrContent, "gumdrop/generator"
-  autoload :GenerationDSL, "gumdrop/generator"
   autoload :HashObject, "gumdrop/hash_object"
   autoload :Pager, "gumdrop/pager"
   autoload :Server, "gumdrop/server"
@@ -27,7 +34,7 @@ module Gumdrop
   
   class << self
     
-    attr_accessor :root_path, :source_path, :site, :layouts, :generators, :partials, :config, :data, :content_filters
+    attr_accessor :root_path, :source_path, :site, :layouts, :generators, :partials, :config, :data, :content_filters, :blacklist
     
     def run(opts={})
       # Opts
@@ -39,7 +46,6 @@ module Gumdrop
       if File.exists? "#{root}/lib/view_helpers.rb"
         # In server mode, we want to reload it every time... right?
         load "#{root}/lib/view_helpers.rb"
-        #require 'view_helpers'
       end
 
       @site        = Hash.new {|h,k| h[k]= nil }
@@ -51,13 +57,12 @@ module Gumdrop
       @data        = Gumdrop::DeferredLoader.new()
 
       @content_filters= []
+      @blacklist      = []
       
       if File.exists? "#{root}/lib/site.rb"
         # In server mode, we want to reload it every time... right?
         source= IO.readlines("#{root}/lib/site.rb").join('')
-        GenerationDSL.class_eval source
-        #load "#{root}/lib/site.rb" 
-        # require 'site' 
+        DSL.class_eval source
       end
       
       # Scan
@@ -89,20 +94,44 @@ module Gumdrop
         generator.execute()
       end
       
+      @blacklist.each do |skip_path|
+        @site.keys.each do |source_path|
+          if source_path.starts_with? skip_path
+            Gumdrop.report " -ignoring: #{source_path}", :info
+            @site.delete source_path
+          end
+        end
+      end
       
       # Render
       unless opts[:dry_run]
         site.keys.sort.each do |path|
-          node= site[path]
-          output_path= "output/#{node.to_s}"
-          FileUtils.mkdir_p File.dirname(output_path)
-          node.renderTo output_path, @content_filters
+          #unless @blacklist.detect {|p| path.starts_with?(p) }
+            node= site[path]
+            output_path= "output/#{node.to_s}"
+            FileUtils.mkdir_p File.dirname(output_path)
+            node.renderTo output_path, @content_filters
+          # else
+          #   Gumdrop.report " -ignoring: #{path}", :info
+          # end
         end
         puts "Done."
       end
       
     end
 
+    # levels: info, warning, error
+    def report(msg, level=:info)
+      ll= Gumdrop.config.log_level
+      case level
+      when :info
+        puts msg if ll == :info
+      when :warning
+        puts msg if ll == :info or ll == :warning
+      else
+        puts msg
+      end
+    end
   end
 
   Gumdrop.config= Gumdrop::HashObject.new(DEFAULT_OPTIONS)